+++ test/nobitex_test.dart
@@ -1,12 +1,30 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
+import 'package:dotenv/dotenv.dart' show load, env;
 
 import 'package:nobitex/nobitex.dart';
 
 void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
+  load();
+
+  final nobitex = Nobitex(token: env['NOBITEX_TOKEN'] as String);
+
+  group('tests for Nobitex', () {
+    test('test getProfile', () async {
+      var data = await nobitex.getProfile();
+
+      expect(data!.containsKey('profile'), true);
+    });
+
+    test('test getWallets', () async {
+      var data = await nobitex.getWallets();
+
+      expect(data!.containsKey('wallets'), true);
+    });
+
+    test('test getWalletRecords', () async {
+      var data = await nobitex.getWalletRecords();
+
+      expect(data!.containsKey('deposits'), true);
+    });
   });
 }