+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
+# Changelog
+## [1.0.0-alpha.1](https://github.com/nekofar/nobitex-api-dart-client/compare/v1.0.0-alpha.0...v1.0.0-alpha.1) (2021-07-15)
 
-* TODO: Describe initial release.
+
+### Features
+
+* add new method to retrieve list of current user wallet records ([d90cd60](https://github.com/nekofar/nobitex-api-dart-client/commit/d90cd603b69f026665a7449df2e5f9ef6ecbfee5))
+
+## 1.0.0-alpha.0 (2021-07-15)
+
+
+### Features
+
+* add new method to receive profile info using api ([8bece56](https://github.com/nekofar/nobitex-api-dart-client/commit/8bece56998e36da610c6ed409c4793061c364ce5))
+* add new method to retrieve list of current user wallets using api ([810cc0c](https://github.com/nekofar/nobitex-api-dart-client/commit/810cc0cd28ac56b87a60d0da58c9a1cfb50636d8))